--- kulprit/data/__init__.py
@@ -1,1 +1,0 @@
-"""Top-level model data module."""
