+++ kulprit/projector.py
@@ -1,0 +1,356 @@
+# pylint: disable=undefined-loop-variable
+# pylint: disable=too-many-instance-attributes
+"""Core reference model class."""
+import warnings
+import numpy as np
+
+from bambi import formula
+from kulprit.plots.plots import plot_compare, plot_densities
+
+from kulprit.projection.arviz_io import compute_loo, get_observed_data, get_pps
+from kulprit.projection.pymc_io import (
+    compile_mllk,
+    compute_llk,
+    compute_new_model,
+    get_model_information,
+)
+from kulprit.projection.search_strategies import user_path, forward_search
+from kulprit.projection.solver import solve
+
+
+class ProjectionPredictive:
+    """
+    Projection Predictive class from which we perform the model selection procedure.
+    """
+
+    def __init__(self, model, idata=None, num_samples=100):
+        """Builder for projection predictive model selection.
+
+        This object initializes the reference model and handles the core projection, variable search
+        methods and submodel comparison.
+
+        Parameters:
+        ----------
+        model : Bambi model
+            The reference GLM model to project
+        idata : InferenceData
+            The ArviZ InferenceData object of the fitted reference model
+        """
+        # get information from Bambi's reference model
+        self.model = model
+        self.has_intercept = formula.formula_has_intercept(self.model.formula.main)
+        self.response_name = self.model.response_component.term.name
+        self.ref_family = self.model.family.name
+        self.priors = self.model.constant_components
+        self.ref_terms = list(
+            self.model.components[self.model.family.likelihood.parent].common_terms.keys()
+        )
+        self.base_terms = self._get_base_terms()
+
+        # get information from PyMC's reference model
+        if not self.model.built:
+            self.model.build()
+        self.pymc_model = self.model.backend.model
+        self.ref_var_info = get_model_information(self.pymc_model)
+        self.all_terms = [fvar.name for fvar in self.pymc_model.free_RVs]
+
+        # get information from ArviZ's InferenceData object
+        self.idata = idata
+        self._check_idata()
+        self.observed_dataset, self.observed_array = get_observed_data(
+            self.idata, self.response_name
+        )
+        self.num_samples = num_samples
+        self.pps = get_pps(self.model, self.idata, self.response_name, self.num_samples)
+        self.elpd_ref = compute_loo(idata=self.idata)
+
+        self.submodels = []
+
+    def __repr__(self) -> str:
+        """Return the terms of the submodels."""
+
+        if not self.submodels:
+            return "ReferenceModel"
+
+        else:
+            str_of_submodels = "\n".join(
+                f"{idx:>3} " f"{value.term_names}" for idx, value in enumerate(self.submodels)
+            )
+            return str_of_submodels
+
+    def project(self, max_terms=None, path="forward"):
+        """Perform model projection.
+
+        If ``max_terms`` is not provided, then the search path runs from the intercept-only model
+        up to but not including the full model.
+
+        Parameters:
+        -----------
+        max_terms : int
+            The number of parameters of the largest submodel in the search path, not including the
+            intercept term.
+        path : str or list
+            The search method to employ, either "forward" for a forward search, or "l1" for
+            a L1-regularized search path. If a nested list of terms is provided, model with
+            those terms will be projected directly.
+        """
+        # test if path is a list of terms
+        if isinstance(path, list):
+            # check if the length of the path always increase
+            if not all(len(path[idx]) < len(path[idx + 1]) for idx in range(len(path) - 1)):
+                raise ValueError("Please provide a list of terms in increasing order")
+            # test if the terms in the path are valid
+            for idx, term_names in enumerate(path):
+                if not set(term_names).issubset(self.all_terms):
+                    raise ValueError(f"Term {idx} is not a valid term in the reference")
+
+            self.submodels = user_path(self._project, path)
+        else:
+            # test valid solution method
+            if path not in ["forward", "l1"]:
+                raise ValueError("Please select either forward search or L1 search.")
+
+            # set default `max_terms` value
+            n_terms = len(self.model.components[self.model.family.likelihood.parent].common_terms)
+            if max_terms is None:
+                max_terms = n_terms
+            # test `max_terms` input
+            elif max_terms > n_terms:
+                warnings.warn(
+                    "max_terms is larger than the number of terms in the reference model."
+                    + "Searching for {n_terms}."
+                )
+                max_terms = n_terms
+
+            if path == "forward":
+                self.submodels = forward_search(self._project, self.ref_terms, max_terms)
+
+            # else:
+            #     self.searcher_path = L1SearchPath(self.projector)
+
+    def _project(self, term_names):
+
+        term_names_ = self.base_terms + term_names
+        new_model = compute_new_model(
+            self.pymc_model, self.ref_var_info, self.all_terms, term_names_
+        )
+
+        model_log_likelihood, old_y_value, obs_rvs = compile_mllk(new_model)
+        initial_guess = np.concatenate(
+            [np.ravel(value) for value in new_model.initial_point().values()]
+        )
+        var_info = get_model_information(new_model)
+
+        new_idata, loss = solve(
+            model_log_likelihood,
+            self.pps,
+            initial_guess,
+            var_info,
+        )
+        # restore obs_rvs value in the model
+        new_model.rvs_to_values[obs_rvs] = old_y_value
+
+        # Add observed data and log-likelihood to the projected InferenceData object
+        new_idata.add_groups(observed_data=self.observed_dataset)
+        new_idata.add_groups(log_likelihood=compute_llk(new_idata, new_model))
+
+        # build SubModel object and return
+        sub_model = SubModel(
+            model=new_model,
+            idata=new_idata,
+            loss=loss,
+            elpd_loo=None,
+            elpd_se=None,
+            size=len(term_names),
+            term_names=term_names,
+            has_intercept=self.has_intercept,
+        )
+        return sub_model
+
+    def _get_base_terms(self):
+        """Extend the model term names to include dispersion terms."""
+
+        base_terms = []
+        # add intercept term if present
+        if self.has_intercept:
+            base_terms.append("Intercept")
+
+        # add the auxiliary parameters
+        if self.priors:
+            aux_params = [f"{str(k)}" for k in self.priors]
+            base_terms += aux_params
+        return base_terms
+
+    def _check_idata(self):
+        # build posterior if not provided
+        if self.idata is None:
+            self.idata = self.model.fit(idata_kwargs={"log_likelihood": True})
+        elif "log_likelihood" not in self.idata.groups():
+            raise UserWarning(
+                """Please run Bambi's fit method with the option
+                idata_kwargs={'log_likelihood': True}"""
+            )
+
+        # test compatibility between model and idata
+        if (
+            not self.model.response_component.term.name
+            == list(self.idata.observed_data.data_vars.variables)[0]
+        ):
+            raise UserWarning("Incompatible model and inference data.")
+
+    def compare(
+        self,
+        plot=True,
+        min_model_size=0,
+        legend=True,
+        title=True,
+        figsize=None,
+        plot_kwargs=None,
+    ):
+        """Compare the ELPD of the projected models along the search path.
+
+
+        Parameters:
+        -----------
+        plot : bool
+            Plot the results of the comparison. Defaults to True
+        legend : bool
+            Add legend to figure. Defaults to True.
+        title : bool
+            Show a tittle with a description of how to interpret the plot. Defaults to True.
+        figsize : tuple
+            If None, size is (10, num of submodels) inches
+        plot_kwargs : dict
+            Optional arguments for plot elements. Currently accepts 'color_elpd', 'marker_elpd',
+        'marker_fc_elpd', 'color_dse', 'marker_dse', 'ls_reference', 'color_ls_reference',
+        'xlabel_rotation'.
+
+        Returns:
+        --------
+        cmp : elpd_info
+            tuples of index, elpd_loo point estimate and standard error for each submodel
+            The index -1 corresponds to the reference model.
+        axes : matplotlib_axes or bokeh_figure
+        """
+        # test that search has been previously run
+        if not self.submodels:
+            raise UserWarning("Please run search before comparing submodels.")
+
+        # initiate plotting arguments if none provided
+        if plot_kwargs is None:
+            plot_kwargs = {}
+
+        elpd_info = [(-1, self.elpd_ref.elpd_loo, self.elpd_ref.se)]
+        # make list with elpd loo and se for each submodel
+        for k, submodel in enumerate(self.submodels):
+            if k >= min_model_size:
+                elpd_info.append((k, submodel.elpd_loo, submodel.elpd_se))
+
+        # plot the comparison if requested
+        axes = None
+        if plot:
+            # list term_names to use  labels
+            if not self.submodels[0].term_names:
+                label_terms = ["Intercept"]
+            else:
+                label_terms = []
+            label_terms.extend(submodel.term_names)
+            label_terms = label_terms[min_model_size:]
+
+            axes = plot_compare(elpd_info, label_terms, legend, title, figsize, plot_kwargs)
+
+        return elpd_info, axes
+
+    def plot_densities(
+        self,
+        var_names=None,
+        submodels=None,
+        include_reference=True,
+        labels="formula",
+        kind="density",
+        figsize=None,
+        plot_kwargs=None,
+    ):
+        """Compare the projected posterior densities of the submodels
+
+        Parameters:
+        -----------
+        var_names : list of str, optional
+            List of variables to plot.
+        submodels : list of int, optional
+            List of submodels to plot, 0 is intercept-only model and the largest valid integer is
+            the total number of variables in reference model. If None, all submodels are plotted.
+        include_reference : bool
+            Whether to include the reference model in the plot. Defaults to True.
+        labels : str
+            If "formula", the labels are the formulas of the submodels. If "size", the number
+            of covariates in the submodels.
+        figsize : tuple
+            Figure size. If None it will be defined automatically.
+        plot_kwargs : dict
+            Dictionary passed to ArviZ's ``plot_density`` function (if kind density) or to
+            ``plot_forest`` (if kind forest).
+
+        Returns:
+        --------
+
+        axes : matplotlib_axes or bokeh_figure
+        """
+        if submodels is None:
+            submodels = self.submodels
+        else:
+            if max(submodels) > len(self.submodels) - 1:
+                submodels = np.clip(submodels, 0, len(self.submodels) - 1)
+                warnings.warn(
+                    "At least one index for the submodels is larger than the number "
+                    "of available submodels"
+                )
+            submodels = [self.submodels[idx] for idx in submodels]
+
+        return plot_densities(
+            self.model,
+            self.idata,
+            var_names=var_names,
+            submodels=submodels,
+            include_reference=include_reference,
+            labels=labels,
+            kind=kind,
+            figsize=figsize,
+            plot_kwargs=plot_kwargs,
+        )
+
+
+class SubModel:
+    """Submodel dataclass.
+
+    Attributes:
+        model (bambi.Model): The submodel's associated Bambi model, from which we can
+            extract a built pymc model.
+        idata (InferenceData): The inference data object of the submodel containing the
+            projected posterior draws and log-likelihood.
+        loo (float): The optimization loss of the submodel
+        size (int): The number of common terms in the model, not including the intercept
+        elpd_loo (float): The expected log pointwise predictive density of the submodel
+        elpd_se (float): The standard error of the expected log pointwise predictive
+        term_names (list): The names of the terms in the model, including the intercept
+        has_intercept (bool): Whether the model has an intercept term
+    """
+
+    def __init__(self, model, idata, loss, size, elpd_loo, elpd_se, term_names, has_intercept):
+        self.model = model
+        self.idata = idata
+        self.loss = loss
+        self.size = size
+        self.elpd_loo = elpd_loo
+        self.elpd_se = elpd_se
+        self.term_names = term_names
+        self.has_intercept = has_intercept
+
+    def __repr__(self) -> str:
+        """String representation of the submodel."""
+        if self.has_intercept:
+            intercept = ["Intercept"]
+        else:
+            intercept = []
+
+        return f"{intercept + self.term_names}"
