--- kulprit/search/__init__.py
@@ -1,12 +1,0 @@
-"""Base search module."""
-
-
-from abc import ABC, abstractmethod
-
-
-class SearchPath(ABC):
-    """Base search module."""
-
-    @abstractmethod
-    def search(self, max_terms: int) -> None:
-        """Initialise search path class."""
